Create cleanCoalData2.m
This was made to performn the same function as cleanCoalData but more efficiently. I need to run them side by side and compare the outputs to make sure it is doing the same thing. This function also includes the functionality of choosing whether emission rates are calculated on a plant level or the asssociated boiler-generator grouping level.